requirements.txt
fastapi
uvicorn
pytubefix
pydub
google-cloud-storage
yt-dlp